requirements/base.txt
Django==6.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
jsonschema-specifications==2025.9.1
PyJWT==2.10.1
python-decouple==3.8
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0
python-dateutil

